requirements.txt
pandas>=2.1
pyyaml>=6.0
pandas_ta>=0.3.14
openai>=1.40.0
matplotlib>=3.7